Tidy up of observer types - move to types package
